requirements.txt
aiogram==3.2.0
google-api-python-client>=2.111.0
google-auth>=2.27.0
aiohttp>=3.9.3
pyqrcode>=1.2.0
python-dotenv==1.0.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0